Refactor P/E/T breadcrumbs implementation using a concern [SCI-8087]
